Update trial3 -All modifiers2.py
Use this when we have to set the modifiers value in a script and then export the human body images to MATLAB for measurement computation.
Follow the steps to export the model as an image:-
1. Open MH--->set White background+Grid off+Show wire mesh+Show smoothened mesh. This has to be the state of model before script is run.
Note: White background scaling will give problems when height is increased, as the humanoid will become longer than the background and then while exporting into MATLAB a perfect binary image will not be created. To solve this issue before starting Makehuman, copy 0_modelling_backgrounds.py file from here and replace it with the same file in Makehuman directory. I have changed the background scaling in this file so that it does not affect the export to MATLAB.
2. After above conditions met run the script the screenshots will be saved. Locate them in Makehuman folder/Grab/  in My documents(Windows) andin Home/Makehuman/v13py/grab/ in Linux.